+++ terraform/slurm_cluster/examples/slurm_cluster/hybrid/simple/main.tf
@@ -68,6 +68,7 @@ locals {
       region             = null
       subnetwork_project = null
       subnetwork         = data.google_compute_subnetwork.default.self_link
+      zone_target_shape  = "ANY_SINGLE_ZONE"
       zone_policy_allow  = []
       zone_policy_deny   = []
     },
@@ -124,6 +125,7 @@ locals {
       region             = null
       subnetwork_project = null
       subnetwork         = data.google_compute_subnetwork.default.self_link
+      zone_target_shape  = "ANY_SINGLE_ZONE"
       zone_policy_allow  = []
       zone_policy_deny   = []
     },
@@ -138,6 +140,12 @@ locals {
     slurm_bin_dir           = "/usr/local/bin"
     output_dir              = "./etc"
     install_dir             = null
+    munge_mount = {
+      server_ip     = null
+      remote_mount  = "/etc/munge"
+      fs_type       = "nfs"
+      mount_options = null
+    }
   }
 }