+++ terraform/slurm_cluster/variables.tf
@@ -331,6 +331,7 @@ EOD
     region             = string
     subnetwork_project = string
     subnetwork         = string
+    zone_target_shape  = string
     zone_policy_allow  = list(string)
     zone_policy_deny   = list(string)
   }))
@@ -384,6 +385,7 @@ EOD
       region             = null
       subnetwork_project = null
       subnetwork         = "default"
+      zone_target_shape  = "ANY_SINGLE_ZONE"
       zone_policy_allow  = []
       zone_policy_deny   = []
     },