+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## \[Next\]
 
 - Add support for custom machine types.
+- Add job id label for exclusive nodes.
 
 ## \[5.5.0\]